new tests on tt's sequence attributes
