+++ testing/bdd/features/validation/sequence_id_num.feature
@@ -11,3 +11,17 @@ Feature: Sequence ID and Sequence Number
         | xml_file            | seq_id   | seq_n |
         | sequence_id_num.xml |          | 5     |
         | sequence_id_num.xml | testSeq1 | a     |
+        | sequence_id_num.xml | testSeq1 | -5    |
+
+    Scenario: Valid Sequence head attributes
+        Given a xml file <xml_file>
+        And it has sequence identifier <seq_id>
+        And it has sequence number <seq_n>
+        Then document is valid
+
+        Examples:
+        | xml_file            | seq_id   | seq_n     |
+        | sequence_id_num.xml | testSeq1 | 5         |
+        | sequence_id_num.xml | a        | 10        |
+        | sequence_id_num.xml | testSeq1 | 999999999 |
+
